Change type of variable i into createChar method: int -> uint8_t.

+++ Deuligne.cpp
@@ -160,7 +160,7 @@ void Deuligne::command( uint8_t command ) {
 void Deuligne::createChar(uint8_t location, uint8_t charmap[]) {
   location &= 0x7; // we only have 8 locations 0-7
   command(LCD_SETCGRAMADDR | (location << 3));
-  for (int i=0; i<8; i++) {
+  for (uint8_t i=0; i<8; i++) {
     write(charmap[i]);
   }
 }
